requirements.txt
torch
torchvision
torchaudio
transformers
easyocr
rapidfuzz
symspellpy
spacy
flask
sentencepiece
pandas
opencv-python
pillow
numpy
accelerate
evaluate
jiwer
scikit-learn
huggingface_hub
fastapi